backend/requirements.txt
absl-py==2.1.0
acres==0.2.0
aiofiles==24.1.0
anyio==4.8.0
asgiref==3.8.1
astunparse==1.6.3
beautifulsoup4==4.13.2
blinker==1.9.0
Bottleneck==1.4.2
build==1.2.2.post1
certifi==2025.1.31
charset-normalizer==3.3.0
chess==1.11.1
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
configobj==5.0.9
configparser==7.1.0
deepface==0.0.93
etelemetry==0.3.1
filelock==3.17.0
fire==0.7.0
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==24.12.23
frontend==0.0.3
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.68.1
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
httplib2==0.22.0
idna==3.10
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
keras==3.8.0
libclang==18.1.1
looseversion==1.3.0
lxml==4.9.3
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
ml-dtypes==0.4.1
mtcnn==1.0.0
namex==0.0.8
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
nltk==3.9.1
numexpr==2.10.2
numpy==1.26.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==11.1.0
pip-tools==7.4.1
protobuf==5.29.3
prov==2.0.1
puremagic==1.28
pydot==3.0.4
Pygments==2.19.1
PyMuPDF==1.25.3
pyparsing==3.2.1
pyproject_hooks==1.2.0
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.1
pyxnat==1.6.2
rdflib==6.3.2
regex==2024.11.6
requests==2.32.3
retina-face==0.0.17
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.1
setuptools==75.8.0
simplejson==3.18.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
starlette==0.45.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tqdm==4.67.1
traits==6.4.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
